# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Integrated Management System API for risks, incidents, audits and CAPA.",
        "title": "IntegraFlow IMS API",
        "contact": {
            "name": "IntegraFlow IMS Team",
            "email": "ims@example.com"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/actions": {
            "get": {
                "description": "Returns actions, optionally filtered by status or source type.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actions"
                ],
                "summary": "List actions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Status filter (Open|In Progress|Done|Overdue)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Source type filter (Risk|Incident|Audit)",
                        "name": "sourceType",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Action"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a corrective/preventive action linked to a risk, incident or audit.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actions"
                ],
                "summary": "Create CAPA action",
                "parameters": [
                    {
                        "description": "Action payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpapi.CreateActionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Action"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/actions/{id}": {
            "put": {
                "description": "Updates the status and/or due date of an action.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actions"
                ],
                "summary": "Update action",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Action ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpapi.UpdateActionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Action"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/audits": {
            "get": {
                "description": "Returns internal audits, optionally filtered by status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audits"
                ],
                "summary": "List audits",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Status filter (Planned|In Progress|Completed)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Audit"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new IMS internal audit record.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audits"
                ],
                "summary": "Create internal audit",
                "parameters": [
                    {
                        "description": "Audit payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpapi.CreateAuditRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Audit"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/audits/{id}": {
            "put": {
                "description": "Updates audit status and/or findings.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audits"
                ],
                "summary": "Update audit",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Audit ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpapi.UpdateAuditRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Audit"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/dashboard": {
            "get": {
                "description": "Returns aggregated IMS KPIs (risks, incidents, actions).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Get IMS dashboard",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Dashboard"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/incidents": {
            "get": {
                "description": "Returns incidents, optionally filtered by domain and status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incidents"
                ],
                "summary": "List incidents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Domain filter",
                        "name": "domain",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Status filter (Open|Investigation|Closed)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Incident"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Records an incident/nonconformity in the IMS.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incidents"
                ],
                "summary": "Create a new incident",
                "parameters": [
                    {
                        "description": "Incident payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpapi.CreateIncidentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Incident"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Action": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dueDate": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "owner": {
                    "type": "string"
                },
                "sourceId": {
                    "type": "integer"
                },
                "sourceType": {
                    "description": "Risk, Incident, Audit",
                    "type": "string"
                },
                "status": {
                    "description": "Open, In Progress, Done, Overdue",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "httpapi.CreateActionRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "dueDate": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "sourceId": {
                    "type": "integer"
                },
                "sourceType": {
                    "description": "risk|incident|audit",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "httpapi.UpdateActionRequest": {
            "type": "object",
            "properties": {
                "dueDate": {
                    "description": "Optional new due date",
                    "type": "string"
                },
                "status": {
                    "description": "Open, In Progress, Done, Overdue",
                    "type": "string"
                }
            }
        },
        "httpapi.UpdateAuditRequest": {
            "type": "object",
            "properties": {
                "findings": {
                    "description": "Summary of audit findings",
                    "type": "string"
                },
                "status": {
                    "description": "Planned, In Progress, Completed",
                    "type": "string"
                }
            }
        },
...
    }
}